+++ SortJson/Functions/ConvertTo-OrderObject.ps1
@@ -1,23 +1,31 @@
 function ConvertTo-OrderObject {
     <#
     .SYNOPSIS
-    Convert a object to a new ordered object
+        Sorts JSON output.
     .DESCRIPTION
-    Convert a object to a new object with the properties sorted in alphabetical order.
-    .PARAMETER Object
-    The object to convert
+        Reformats a JSON string so the output looks better than what ConvertTo-Json outputs.
+    .PARAMETER Json
+        Required: [string] The JSON text to sort.
+    .PARAMETER ReSorted
+        Optional: ReSorted the json properties to the Sorted specified in the PropertyStartList and PropertyEndList parameters.
     .PARAMETER PropertyStartList
-    The list of properties to put at the start of the object
+        Required: The list of properties to put at the start of the json object.
     .PARAMETER PropertyEndList
-    The list of properties to put at the end of the object
+        Required: The list of properties to put at the end of the json object.
     .PARAMETER OnlyListedProperties
-    Only include the properties listed in the PropertyStartList and PropertyEndList
+        Optional: Only list the properties that are in the PropertyStartList and PropertyEndList.
     .PARAMETER SortAlphabetically
-    Sort the properties in alphabetical order
-    .PARAMETER Recursive
-    Sort the properties that are arrays or objects
+        Optional: Sort the properties alphabetically.
     .EXAMPLE
-    ConvertTo-OrderObject -Object $JsonObj -PropertyStartList @("Name","Age") -PropertyEndList @("Address","Phone")
+        $json | ConvertFrom-Json | ConvertTo-OrderObject -PropertyStartList @('displayName','name','description','version','publisher') -PropertyEndList @('settings','assignments') -SortAlphabetically -recursive
+    .EXAMPLE
+        $PropertyStartList = $script:SortedProperties.firstOrder
+        $PropertyEndList = $script:SortedProperties.lastOrder
+        $json | ConvertFrom-Json | ConvertTo-OrderObject -PropertyStartList $PropertyStartList -PropertyEndList $PropertyEndList -SortAlphabetically
+    .NOTES
+        https://stackoverflow.com/questions/56322993/proper-formating-of-json-using-powershell
+    .LINK
+        Set-ObjectPropertyOrder
     #>
     [CmdletBinding()]
     Param(
@@ -33,6 +41,8 @@ function ConvertTo-OrderObject {
         [switch]$OnlyListedProperties,
 
         [switch]$SortAlphabetically,
+
+        [switch]$IgnoreCaseSensitivity,
 
         [switch]$Recursive
     )
@@ -59,7 +69,7 @@ function ConvertTo-OrderObject {
 
                 Write-Verbose ("{0} :: Property Start List: {1}" -f ${CmdletName}, ($NewPropertyStartList -join ", "))
             }
-            
+
             If($PropertyEndList.Count -gt 0){
                 $NewPropertyEndList = @()
 
@@ -78,7 +88,7 @@ function ConvertTo-OrderObject {
                 $RemainingProperties = ($JsonObj | Select-Object -ExcludeProperty ($NewPropertyStartList + $NewPropertyEndList)).PSObject.Properties.Name
                 Write-Verbose ("{0} :: Remaining properties found: {1}" -f ${CmdletName}, ($RemainingProperties -join ", "))
             }
-            
+
 
             #build a new object in the Sorted we want
             $SortedObj = New-Object PSObject
@@ -94,18 +104,18 @@ function ConvertTo-OrderObject {
                 {
                     $f++
                     Write-Verbose ("{0} :: Adding first property set [{1} of {2}]: {3}" -f ${CmdletName}, $f,$NewPropertyEndList.count,$Property)
-                    
+
                     #add the property to the new object
                     $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonObj.$Property -Force
 
                     #recursivly call the function to sort the properties that are arrays or objects, then update the property
                     If( ($Recursive -eq $true) -and ($SortAlphabetically -eq $true) )
                     {
-                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property
+                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property -IgnoreCaseSensitivity:$IgnoreCaseSensitivity
 
                         Write-Verbose ("{0} :: Adding property value: {1}" -f ${CmdletName}, $JsonProperty.Tostring())
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonProperty -Force
-                    
+
                     }Else{
                         Write-Verbose ("{0} :: Adding remaining property: {1}" -f ${CmdletName}, $Property)
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonObj.$Property -Force
@@ -113,14 +123,14 @@ function ConvertTo-OrderObject {
                 }
             }
 
-            
+
             If( $OnlyListedProperties -ne $true )
             {
                 #add the properties that are not in the start or end list
                 $PropertyMiddleList = $JsonObj | Select-Object -ExcludeProperty ($PropertyStartList + $PropertyEndList)
-                
+
                 If($SortAlphabetically -eq $true){
-                    $PropertyMiddleList = $PropertyMiddleList.PSObject.Properties.Name | Sort-Object
+                    $PropertyMiddleList = $PropertyMiddleList.PSObject.Properties.Name | Sort-Object -CaseSensitive:(!$IgnoreCaseSensitivity)
                 }Else{
                     $PropertyMiddleList = $PropertyMiddleList.PSObject.Properties.Name
                 }
@@ -130,57 +140,57 @@ function ConvertTo-OrderObject {
                     #recursivly call the function to sort the properties that are arrays or objects, then update the property
                     If( ($Recursive -eq $true) -and ($SortAlphabetically -eq $true) )
                     {
-                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property
+                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property -IgnoreCaseSensitivity:$IgnoreCaseSensitivity
 
                         Write-Verbose ("{0} :: Adding property value: {1}" -f ${CmdletName}, $JsonProperty.Tostring())
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonProperty -Force
-                    
+
                     }Else{
                         Write-Verbose ("{0} :: Adding remaining property: {1}" -f ${CmdletName}, $Property)
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonObj.$Property -Force
                     }
                 }
-                
+
             }
-            
+
             #add the end properties
             If($NewPropertyEndList.count -gt 0)
-            {                
+            {
                 $l=0
                 Foreach($Property in $NewPropertyEndList)
                 {
                     $l++
                     Write-Verbose ("{0} :: Adding last property set [{1} of {2}]: {3}" -f ${CmdletName}, $l,$NewPropertyEndList.count,$Property)
-                    
+
                     #add the property to the new object
                     $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonObj.$Property -Force
 
                     #recursivly call the function to sort the properties that are arrays or objects, then update the property
                     If( ($Recursive -eq $true) -and ($SortAlphabetically -eq $true) )
                     {
-                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property
+                        $JsonProperty = Set-ObjectPropertyOrder -Object $JsonObj -Property $Property -IgnoreCaseSensitivity:$IgnoreCaseSensitivity
 
                         Write-Verbose ("{0} :: Adding property value: {1}" -f ${CmdletName}, $JsonProperty.Tostring())
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonProperty -Force
-                    
+
                     }Else{
                         Write-Verbose ("{0} :: Adding remaining property: {1}" -f ${CmdletName}, $Property)
                         $SortedObj | Add-Member -MemberType NoteProperty -Name $Property -Value $JsonObj.$Property -Force
                     }
                 }
-    
+
             }
         }
-        
+
         #add the sorted json to the list
         #$ObjList += ($SortedObj | ConvertTo-Json) -replace '\\"','"' -replace '\\r\\n','' -replace '"{','{' -replace '}"','}'
         Write-Verbose ("{0} :: Collecting objects..." -f ${CmdletName})
         $ObjList += $SortedObj
-        #$ObjList += ($SortedObj | ConvertTo-Json) 
+        #$ObjList += ($SortedObj | ConvertTo-Json)
     }
     End{
         Write-Verbose ("{0} :: Returning object..." -f ${CmdletName})
         return $ObjList
     }
-    
+
 }